main/requirements.txt
Flask[async]==2.3.2
PyJWT[crypto]==2.7.*
httpx==0.24.1
